+++ include/core/managers/EnemyManager.hpp
@@ -10,4 +10,5 @@ public:
     ~EnemyManager();
 
     void SpawnEnemy(const FVector2& position, const FVector2& size, const Color color, float accelerationIndex);
+    const std::vector<std::weak_ptr<GameObject>>& GetManagedObjects() const { return _ManagedObjects; }
 };