+++ include/spaceInvaders/Bullet.hpp
@@ -7,7 +7,7 @@ class Bullet: public Circle, public std::enable_shared_from_this<Bullet>{
 public:
     std::function<void(std::shared_ptr<Bullet>)> OnExpired;
     Bullet(Game* game, GameObject* owner, FVector2 position = FVector2{0,0}, float radius = 10.0f
-        , Color color = RAYWHITE, float accelerationIndex = 200.0f);
+        , Color color = RAYWHITE, float accelerationIndex = 200.0f, FVector2 velocity = FVector2{0, -1});
     ~Bullet();
 
     virtual void Update(float deltaTime) override;